dont download chrome bin if its already there compress and uglify compiled source

+++ buildCi.sh
@@ -5,29 +5,38 @@
 #
 
 
-# 
+#
 # varables
-# 
+#
 chromefile="chromium.zip"
+chromebin="headless-chromium"
 zipfile="lambda.zip"
 
-# 
+#
 # compile
-# 
+#
 npm run clean
-curl -L https://github.com/adieuadieu/serverless-chrome/releases/download/v1.0.0-55/stable-headless-chromium-amazonlinux-2017-03.zip -o $chromefile
-unzip $chromefile
+if [ ! -f "$chromebin" ]; then
+  echo "chome bin not found... downloading"
+  curl -L https://github.com/adieuadieu/serverless-chrome/releases/download/v1.0.0-55/stable-headless-chromium-amazonlinux-2017-03.zip -o $chromefile
+  unzip $chromefile
+  
+fi
 export PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
 export NODE_ENV=production
 npm ci --only=production
 # npm -g install typescript
 npx -p typescript tsc
+for jsfile in "build"/*.js
+do
+  npx uglify-es --compress --mangle -o ./"$jsfile" -- ./"$jsfile"
+done
 
 #
 # package
-# 
+#
 zip -r $zipfile node_modules
 cd ./build
-zip -ru ../$zipfile .
+zip -ru ../$zipfile *.js
 cd ..
 zip -ju $zipfile ./headless-chromium
